refactor: colocate uniforms json schema utils and other form related file
