+++ libs/frontend/domain/atom/src/use-cases/update-atom/updateAtomSchema.ts
@@ -1,7 +1,7 @@
 import type { IUpdateAtomDTO } from '@codelab/frontend/abstract/core'
 import { filterNotHookType } from '@codelab/frontend/abstract/core'
+import { nonEmptyString } from '@codelab/frontend/view/components'
 import { IAtomType } from '@codelab/shared/abstract/core'
-import { nonEmptyString } from '@codelab/shared/utils'
 import type { JSONSchemaType } from 'ajv'
 
 export const updateAtomSchema: JSONSchemaType<IUpdateAtomDTO> = {
