+++ libs/frontend/domain/component/src/use-cases/update-component/updateComponentSchema.ts
@@ -1,7 +1,7 @@
 import type { IUpdateComponentDTO } from '@codelab/frontend/abstract/core'
 import { getSelectElementComponent } from '@codelab/frontend/domain/type'
+import { nonEmptyString } from '@codelab/frontend/view/components'
 import { ElementTypeKind } from '@codelab/shared/abstract/codegen'
-import { nonEmptyString } from '@codelab/shared/utils'
 import type { JSONSchemaType } from 'ajv'
 
 export const updateComponentSchema: JSONSchemaType<IUpdateComponentDTO> = {
