+++ libs/frontend/domain/element/src/use-cases/element/create-element/createElementSchema.ts
@@ -1,6 +1,6 @@
 import type { ICreateElementDTO } from '@codelab/frontend/abstract/core'
 import { RenderTypeEnum } from '@codelab/frontend/abstract/core'
-import { titleCaseValidation } from '@codelab/shared/utils'
+import { titleCaseValidation } from '@codelab/frontend/view/components'
 import type { JSONSchemaType } from 'ajv'
 
 export const createElementSchema: JSONSchemaType<ICreateElementDTO> = {
