+++ libs/frontend/domain/element/src/use-cases/element/update-element/updateElementSchema.ts
@@ -1,6 +1,6 @@
 import type { IUpdateBaseElementDTO } from '@codelab/frontend/abstract/core'
 import { RenderTypeEnum } from '@codelab/frontend/abstract/core'
-import { titleCaseValidation } from '@codelab/shared/utils'
+import { titleCaseValidation } from '@codelab/frontend/view/components'
 import type { JSONSchemaType } from 'ajv'
 
 export const updateElementSchema: JSONSchemaType<IUpdateBaseElementDTO> = {
