+++ libs/frontend/domain/store/src/use-cases/actions/create-action/createActionSchema.ts
@@ -3,14 +3,12 @@ import { HttpMethod, HttpResponseType } from '@codelab/frontend/abstract/core'
 import {
   CodeMirrorField,
   CodeMirrorGraphqlField,
-} from '@codelab/frontend/view/components'
-import { CodeMirrorLanguage } from '@codelab/shared/abstract/codegen'
-import { IActionKind } from '@codelab/shared/abstract/core'
-import {
   hideField,
   nonEmptyString,
   showFieldOnDev,
-} from '@codelab/shared/utils'
+} from '@codelab/frontend/view/components'
+import { CodeMirrorLanguage } from '@codelab/shared/abstract/codegen'
+import { IActionKind } from '@codelab/shared/abstract/core'
 import type { JSONSchemaType } from 'ajv'
 import keys from 'lodash/keys'
