+++ libs/frontend/view/components/form/index.ts
@@ -6,4 +6,4 @@ export * from './hooks/submitController'
 export * from './hooks/uniformUtils'
 export * from './modal'
 export * from './providers'
-export * from './schema/emptyJsonSchema'
+export * from './schema'
