+++ libs/frontend/view/components/form/schema/hideField.ts
@@ -1,0 +1,5 @@
+export const hideField = {
+  uniforms: {
+    component: () => null,
+  },
+}
