+++ libs/frontend/view/components/form/schema/idField.ts
@@ -1,0 +1,9 @@
+import { hideField } from './hideField'
+
+export const idField = {
+  id: {
+    type: 'string',
+    nullable: true,
+    ...hideField,
+  },
+}
