+++ libs/frontend/view/components/form/schema/nonEmptyString.ts
@@ -1,0 +1,5 @@
+export const nonEmptyString = {
+  type: 'string' as const,
+  transform: ['trim'],
+  minLength: 1,
+}
