+++ libs/frontend/view/components/form/schema/showFieldOnDev.ts
@@ -1,0 +1,11 @@
+export const showFieldOnDev = () => {
+  if (process.env.NODE_ENV === 'development') {
+    return {}
+  }
+
+  return {
+    uniforms: {
+      component: () => null,
+    },
+  }
+}
