+++ libs/frontend/view/components/form/schema/titleCaseValidation.ts
@@ -1,0 +1,6 @@
+import { singlySpacedTitleCaseWithNumbersRegex } from '@codelab/shared/utils'
+
+export const titleCaseValidation = {
+  pattern: singlySpacedTitleCaseWithNumbersRegex.source,
+  errorMessage: 'must be in Title Case',
+}
