+++ libs/shared/utils/src/index.ts
@@ -6,4 +6,3 @@ export * from './mapDeep'
 export * from './props'
 export * from './regex'
 export * from './transform'
-export * from './uniform'
